backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.0.1"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "~0.11.3"
fastapi = "~0.68.0"
python-multipart = "~0.0.5"
pydantic = {extras = ["email"], version = "~1.8.1"}
Jinja2 = "3.0.1"
SQLAlchemy = "^1.4.22"
alembic = "^1.6.5"
tenacity = "^8.0.1"
greenlet = "^1.1.2"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
cffi = "^1.15.0"